Split Registrar into Static and Dynamic variants
