+++ Libs/GObject-2.0/Native/Classes/TypeRegistrar.Static.cs
@@ -1,0 +1,15 @@
+﻿namespace GObject.Native
+{
+    public class StaticRegistrar : TypeRegistrar
+    {
+        public void RegisterSubclassStatic(System.Type type, System.Type parentType, string qualifiedName)
+        {
+            // Do actual registration
+            RegisterSubclass(
+                type,
+                parentType,
+                qualifiedName 
+            );
+        }
+    }
+}
