+++ Libs/GObject-2.0/Native/Classes/TypeRegistrar.cs
@@ -8,15 +8,9 @@ namespace GObject.Native
     {
         public TypeRegistrationException(string message) : base(message) { }
     }
-    
-    internal static class TypeRegistrar
-    {
-        // TODO: Split into two classes + abstract base class so we can
-        // switch between reflection and source generation at compile time
 
-        private static string QualifyName(System.Type type)
-            => type.FullName?.Replace(".", "") ?? type.Name;
-        
+    public abstract class TypeRegistrar
+    {
         private static TypeQuery.Struct GetTypeMetrics(Type parentType)
         {
             TypeQuery.Handle handle = TypeQuery.ManagedHandle.Create();
@@ -24,23 +18,9 @@ namespace GObject.Native
 
             return Marshal.PtrToStructure<TypeQuery.Struct>(handle.DangerousGetHandle());
         }
-
-        public static void RegisterSubclass(System.Type type)
-        {
-            Debug.Assert(
-                condition: type.BaseType != null,
-                message: "Cannot register a toplevel type - it must inherit from at least GObject"
-            );
-            
-            System.Type baseType = type.BaseType;
-            if (!TypeDictionary.ContainsSystemType(baseType))
-                RegisterSubclass(baseType);
-            
-            RegisterType(type);
-        }
         
         // Registers a new type class with the underlying GType type system
-        private static void RegisterType(System.Type type)
+        protected void RegisterSubclass(System.Type type, System.Type parentType, string qualifiedName)
         {
             Debug.Assert(
                 condition: !TypeDictionary.ContainsSystemType(type),
@@ -53,8 +33,8 @@ namespace GObject.Native
             );
             
             // Get metrics about parent type
-            Type parentType = TypeDictionary.GetGType(type.BaseType!);
-            TypeQuery.Struct query = GetTypeMetrics(parentType);
+            Type parentGType = TypeDictionary.GetGType(parentType);
+            TypeQuery.Struct query = GetTypeMetrics(parentGType);
 
             if (query.Type == 0)
                 throw new TypeRegistrationException("Could not query parent type");
@@ -69,11 +49,10 @@ namespace GObject.Native
             };
 
             // Perform Registration
-            var qualifiedName = QualifyName(type);
             Console.WriteLine($"Registering type {type.Name} as {qualifiedName}");
 
             TypeInfo.Handle handle = TypeInfo.ManagedHandle.Create(typeInfo);
-            var typeid = Functions.TypeRegisterStatic(parentType.Value, qualifiedName, handle, 0);
+            var typeid = Functions.TypeRegisterStatic(parentGType.Value, qualifiedName, handle, 0);
             
             if (typeid == 0)
                 throw new TypeRegistrationException("Type Registration Failed!");
